Fix an issue with the reader returning 0, nil
While this was exected behaivor and defined in the docs, it is
non-standard everywhere else in Go land. For that reason and because a
bunch of different libraries depend on that readers behaving that way I
am only returning on Read when we have some data to give. Thus, the new
expected behavior is that calling follower.Read(b) will block until data
can be returned.